Add new blobstore backend using Weed-FS
The main package is at pkg/blobserver/weedfs - modeled after pkg/blobserver/s3.

It uses some helpers for communicating with the Weed-FS master, in package
third_party/github.com/tgulacsi/weedfs-client. As this is only one http.go file,
maybe this could be incorporated in the weedfs package, as well.

Change-Id: I7940d47faabe63efb75c9e22b2d7715347092890