Diskpacked: move to (short) JSON-based header format
Add repack-diskpacked command to camtool, as for the diskpacked format
change, a repack mechanism is useful (not strictly needed, as we still can
read the old format, and index it, too!).

This format starts every diskpacked pack file with a two byte version tag
("V1"), and every header with "\nDISKPCK" bytes, following the header
length (one byte) followed by the JSON-encoded header.

A version-sniffer and a version-based pack file walker is introduced,
to be able to walk (and possibly convert) old files, too.

If the last pack file has an older version, then a new pack file is created.

Change-Id: I805ffe908e38d32ec5cd473514002df17ca01259